Module rlifesrc_lib::config

World configuration.

Structs

Config
World configuration.

Enums

NewState
How to choose a state for an unknown cell.

SearchOrder
The order to find a new unknown cell.

SkipLevel
What patterns are considered boring and should be skip.

Symmetry
Symmetries of the pattern.

Transform
Transformations (rotations and reflections) after the last generation in a period.