Enum rlifesrc_lib::poly_world::PolyWorld

#[non_exhaustive]
pub enum PolyWorld {
    Life(World<Life, LifeSrc>),
    LifeGen(World<LifeGen, LifeSrc>),
    NtLife(World<NtLife, LifeSrc>),
    NtLifeGen(World<NtLifeGen, LifeSrc>),
    LifeBackjump(World<Life, Backjump<Life>>),
    NtLifeBackjump(World<NtLife, Backjump<NtLife>>),
}

A polymorphic World.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Life(World<Life, LifeSrc>)

A World with Life rule and LifeSrc algorithm.

LifeGen(World<LifeGen, LifeSrc>)

A World with LifeGen rule and LifeSrc algorithm.

NtLife(World<NtLife, LifeSrc>)

A World with NtLife rule and LifeSrc algorithm.

NtLifeGen(World<NtLifeGen, LifeSrc>)

A World with NtLifeGen rule and LifeSrc algorithm.

LifeBackjump(World<Life, Backjump<Life>>)

A World with Life rule and Backjump algorithm.

NtLifeBackjump(World<NtLife, Backjump<NtLife>>)

A World with NtLife rule and Backjump algorithm.

Implementations

impl PolyWorld

pub fn search(&mut self, max_step: Option<u64>) -> Status

The search function.

Returns Status::Found if a result is found, Status::None if such pattern does not exist, Status::Searching if the number of steps exceeds max_step and no results are found.

pub fn get_cell_state(&self, coord: Coord) -> Option<State>

Gets the state of a cell. Returns Err(()) if there is no such cell.

pub const fn config(&self) -> &Config

World configuration.

pub const fn is_gen_rule(&self) -> bool

Whether the rule is a Generations rule.

pub fn is_b0_rule(&self) -> bool

Whether the rule contains B0.

In other words, whether a cell would become ALIVE in the next generation, if all its neighbors in this generation are dead.

pub fn cell_count_gen(&self, t: i32) -> u32

Number of known living cells in some generation.

For Generations rules, dying cells are not counted.

pub fn cell_count(&self) -> u32

Minimum number of known living cells in all generation.

For Generations rules, dying cells are not counted.

pub const fn conflicts(&self) -> u64

Number of conflicts during the search.

pub fn set_max_cell_count(&mut self, max_cell_count: Option<u32>)

Set the max cell counts.

Currently this is the only parameter that you can change during the search.

pub fn rle_gen(&self, t: i32) -> String

Displays the whole world in some generation, in a mix of Plaintext and RLE format.

Dead cells are represented by .;

Living cells are represented by o for rules with 2 states, A for rules with more states;

Dying cells are represented by uppercase letters starting from B;

Unknown cells are represented by ?;

Each line is ended with $;

The whole pattern is ended with !.

pub fn plaintext_gen(&self, t: i32) -> String

Displays the whole world in some generation in Plaintext format.

Do not use this for Generations rules.

Dead cells are represented by .;

Living and Dying cells are represented by o;

Unknown cells are represented by ?.

pub fn ser(&self) -> WorldSer

Available on crate feature serde only.

Saves the world as a WorldSer, which can be easily serialized.

pub fn deser(&mut self, ser: &WorldSer) -> Result<(), Error>

Available on crate feature serde only.

Restores the world from the WorldSer.