Module rlifesrc_lib::rules

Cellular automata rules.

For the notations of rule strings, please see this article on LifeWiki.

Modules

life
Totalistic Life-like rules.

macros
A macro to generate the corresponding Generations rule of a rule.

ntlife
Non-totalistic Life-like rules.

Structs

Life
Totalistic Life-like rules.

LifeGen
Totalistic Life-like Generations rules.

NtLife
Non-totalistic Life-like rules.

NtLifeGen
Non-totalistic Life-like Generations rules.

Traits

Rule
A cellular automaton rule.