Struct rlifesrc_lib::rules::NtLife

pub struct NtLife {
    b0: bool,
    s8: bool,
    symmetry: Symmetry,
    impl_table: Vec<ImplFlags>,
}

Non-totalistic Life-like rules.

This includes any rule that can be converted to a non-totalistic Life-like rule: isotropic non-totalistic rules, non-isotropic rules, hexagonal rules, rules with von Neumann neighborhoods, etc.

Fields
b0: bool

Whether the rule contains B0.

s8: bool

Whether the rule contains S8.

symmetry: Symmetry

The symmetry of the rule.

impl_table: Vec<ImplFlags>

An array of actions for all neighborhood descriptors.

Implementations

impl NtLife

pub fn new(b: &[u8], s: &[u8]) -> Self

Constructs a new rule from the b and s data.

fn init_symmetry(self, b: &[u8], s: &[u8]) -> Self

Deduces the symmetry of the rule

fn init_trans(self, b: &[u8], s: &[u8]) -> Self

Deduces the implication for the successor.

fn init_conflict(self) -> Self

Deduces the conflicts.

fn init_impl(self) -> Self

Deduces the implication for the cell itself.

fn init_impl_nbhd(self) -> Self

Deduces the implication for the neighbors.