Struct rlifesrc_lib::rules::NtLifeGen

pub struct NtLifeGen {
    b0: bool,
    s8: bool,
    gen: usize,
    symmetry: Symmetry,
    impl_table: Vec<ImplFlags>,
}

Non-totalistic Life-like Generations rules.

This includes any rule that can be converted to a non-totalistic Life-like Generations rule.

Fields
b0: bool

Whether the rule contains B0.

s8: bool

Whether the rule contains S8.

gen: usize

Number of states.

symmetry: Symmetry

The symmetry of the rule.

impl_table: Vec<ImplFlags>

An array of actions for all neighborhood descriptors.

Implementations

impl NtLifeGen

pub fn new(b: &[u8], s: &[u8], gen: usize) -> Self

Constructs a new rule from the b and s data and the number of states.

pub fn non_gen(self) -> NtLife

Converts to the corresponding non-Generations rule.