Trait rlifesrc_lib::rules::Rule

pub trait Rule: Sealed {
    type Desc: Copy;

    const IS_GEN: bool;

    fn has_b0(&self) -> bool;

    fn has_b0_s8(&self) -> bool;

    fn gen(&self) -> usize;
}

A cellular automaton rule.

Some details of this trait is hidden in the doc. Please use the following structs instead of implementing by yourself:

Life

LifeGen

NtLife

NtLifeGen

Associated Types

type Desc: Copy

The type of neighborhood descriptor of the rule.

It describes the states of the successor and neighbors of a cell, and is used to determine the state of the cell in the next generation.

Loading content...
Associated Constants

const IS_GEN: bool

Whether the rule is a Generations rule.

Loading content...
Required methods

fn has_b0(&self) -> bool

Whether the rule contains B0.

In other words, whether a dead cell would become ALIVE in the next generation, if all its neighbors in this generation are dead.

fn has_b0_s8(&self) -> bool

Whether the rule contains both B0 and S8.

In a rule that contains B0, a dead cell would become ALIVE in the next generation, if all its neighbors in this generation are dead.

In a rule that contains S8, a living cell would stay ALIVE in the next generation, if all its neighbors in this generation are alive.

fn gen(&self) -> usize

The number of states.

Loading content...