Trait rlifesrc_lib::search::Algorithm

pub trait Algorithm<'a, R: Rule>: Sealed {
    type Reason: Reason<'a, R>;
    fn new() -> Self;

    fn deser_reason(
        world: &World<'a, R, Self>, 
        ser: &ReasonSer
    ) -> Result<Self::Reason, Error>;
}

The search algorithms.

Currently only two algorithms are supported:

LifeSrc: The default algorithm based on David Bell’s lifesrc.

Backjump: (Experimental) Adding Backjumping to the original lifesrc algorithm. Very slow. Do not use it.

Some details of this trait is hidden in the doc. Please use the above structs instead of implementing by yourself.

Associated Types

type Reason: Reason<'a, R>

Reasons for setting a cell.

Required methods

fn new() -> Self

Generate new algorithm data.

fn deser_reason(
world: &World<'a, R, Self>,
ser: &ReasonSer
) -> Result<Self::Reason, Error>

This is supported on crate feature serde only.

Restore the reason from a ReasonSer.