Struct rlifesrc_lib::Config

pub struct Config {
    pub width: isize,
    pub height: isize,
    pub period: isize,
    pub dx: isize,
    pub dy: isize,
    pub transform: Transform,
    pub symmetry: Symmetry,
    pub search_order: Option<SearchOrder>,
    pub new_state: NewState,
    pub max_cell_count: Option<usize>,
    pub non_empty_front: bool,
    pub reduce_max: bool,
    pub rule_string: String,
    pub diagonal_width: Option<isize>,
    pub skip_level: SkipLevel,
}

World configuration.

The world will be generated from this configuration.

Fields
width: isize

Width.

height: isize

Height.

period: isize

Period.

dx: isize

Horizontal translation.

dy: isize

Vertical translation.

transform: Transform

Transformations (rotations and reflections) after the last generation in a period.

After the last generation in a period, the pattern will return to the first generation, applying this transformation first, and then the translation defined by dx and dy.

symmetry: Symmetry

Symmetries of the pattern.

search_order: Option<SearchOrder>

The order to find a new unknown cell.

It will always search all generations of one cell before going to another cell.

None means that it will automatically choose a search order according to the width and height of the world.

new_state: NewState

How to choose a state for an unknown cell.

max_cell_count: Option<usize>

The number of minimum living cells in all generations must not exceed this number.

None means that there is no limit for the cell count.

non_empty_front: bool

Whether to force the first row/column to be nonempty.

Depending on the search order, the 'front' means:

the first row, when the search order is row first;

the first column, when the search order is column first;

the first row plus the first column, when the search order is diagonal.

reduce_max: bool

Whether to automatically reduce the max_cell_count when a result is found.

The max_cell_count will be set to the cell count of the current result minus one.

rule_string: String

The rule string of the cellular automaton.

diagonal_width: Option<isize>

Diagonal width.

If the diagonal width is n, the cells at position (x, y) where abs(x - y) >= n are assumed to be dead.

skip_level: SkipLevel

What patterns are considered boring and should be skip.

Implementations

impl Config

pub fn new(width: isize, height: isize, period: isize) -> Self

Sets up a new configuration with given size.

pub fn set_translate(self, dx: isize, dy: isize) -> Self

Sets the translations (dx, dy).

pub fn set_transform(self, transform: Transform) -> Self

Sets the transformation.

pub fn set_symmetry(self, symmetry: Symmetry) -> Self

Sets the symmetry.

pub fn set_search_order<T: Into<Option<SearchOrder>>>(
self,
search_order: T
) -> Self

Sets the search order.

pub fn set_new_state(self, new_state: NewState) -> Self

Sets how to choose a state for an unknown cell.

pub fn set_max_cell_count<T: Into<Option<usize>>>(
self,
max_cell_count: T
) -> Self

Sets the maximal number of living cells.

pub fn set_non_empty_front(self, non_empty_front: bool) -> Self

Sets whether to force the first row/column to be nonempty.

pub fn set_reduce_max(self, reduce_max: bool) -> Self

Sets whether to automatically reduce the max_cell_count when a result is found.

pub fn set_rule_string<S: ToString>(self, rule_string: S) -> Self

Sets the rule string.

pub fn set_diagonal_width<T: Into<Option<isize>>>(
self,
diagonal_width: T
) -> Self

Sets the diagonal width.

pub fn set_skip_level(self, skip_level: SkipLevel) -> Self

Sets the skip level.

pub(crate) fn auto_search_order(&self) -> SearchOrder

Automatically determines the search order if search_order is None.

pub(crate) fn search_order_iter(&self) -> Box<dyn Iterator<Item = Coord>>

Generates an iterator over cells coordinates from the search order.

pub(crate) fn translate(&self, coord: Coord) -> Coord

Applies the transformation and translation to a coord.

pub fn world(&self) -> Result<Box<dyn Search>, Error>

Creates a new world from the configuration. Returns an error if the rule string is invalid.