Module rlifesrc_lib::config

World configuration.

Re-exports

pub use d8::Symmetry;

pub use d8::Transform;

pub use search_order::SearchOrder;

Modules

d8

Configurations related to the dihedral group D8.

search_order

Configurations related to the the search order.

Structs

Config

World configuration.

KnownCell

A cell whose state is known before the search.

Enums

NewState

How to choose a state for an unknown cell.