Trait rlifesrc_lib::search::Reason

pub trait Reason<'a, R: Rule>: Sized {
    type ConflReason;

    const KNOWN: Self;
    const DECIDED: Self;

    fn from_cell(cell: CellRef<'a, R>) -> Self;

    fn from_sym(cell: CellRef<'a, R>) -> Self;

    fn is_decided(&self) -> bool;

    fn ser(&self) -> ReasonSer;

    fn deser(ser: &ReasonSer, world: &World<'a, R, Self>) -> Result<Self, Error>;
}

Reasons for setting a cell.

Different choices of reason set will result in different choices of algorithm.

Some details of this trait is hidden in the doc. Please use the following structs instead of implementing by yourself:

ReasonBackjump

ReasonNoBackjump

Associated Types

type ConflReason

Reasons for a conflict.

Loading content...
Associated Constants

const KNOWN: Self

Known before the search starts,

const DECIDED: Self

Decides the state of a cell by choice.

Loading content...
Required methods

fn from_cell(cell: CellRef<'a, R>) -> Self

Deduced from the rule when constitifying another cell.

fn from_sym(cell: CellRef<'a, R>) -> Self

Deduced from symmetry.

fn is_decided(&self) -> bool

Decided or trying another state for generations rules.

fn ser(&self) -> ReasonSer

Saves the reason as a ReasonSer.

fn deser(ser: &ReasonSer, world: &World<'a, R, Self>) -> Result<Self, Error>

Restore the reason from a ReasonSer.

Loading content...